# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/home/detalhes/{id}": {
      "get": {
        "operationId": "TasksController_getTaskById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/home/criar-tarefa": {
      "post": {
        "operationId": "TasksController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/home/buscar/{search}": {
      "post": {
        "operationId": "TasksController_searchTask",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/home/{id}": {
      "patch": {
        "operationId": "TasksController_updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "NestJS Intro",
    "description": "API para gerenciamento projeto pessoal",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "titleTask": {
            "type": "string"
          },
          "descriptionTask": {
            "type": "string"
          }
        },
        "required": [
          "titleTask",
          "descriptionTask"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "string"
          },
          "titleTask": {
            "type": "string"
          },
          "descriptionTask": {
            "type": "string"
          },
          "idUser": {
            "type": "string"
          },
          "idTask": {
            "type": "string"
          }
        },
        "required": [
          "completed",
          "titleTask",
          "descriptionTask",
          "idUser",
          "idTask"
        ]
      },
...
    }
  }
}